% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{au_2006}
\alias{au_2006}
\title{Australia Time Use Acitivity Classification}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 224 rows and 4 columns.
}
\source{
\url{https://www.abs.gov.au/AUSSTATS/subscriber.nsf/log?openagent&41500\%20tus\%202006\%20curf\%20data\%20items.xls&4150.0&Data\%20Cubes&87316ADE9E373C4CCA2573F500151342&0&2006&21.02.2008&Latest}
}
\usage{
au_2006
}
\description{
See \url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/Latestproducts/863BE05F5DF81C08CA2573F500148972?opendocument}
}
\keyword{datasets}
