% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icatus_2016}
\alias{icatus_2016}
\title{International Classification of Activities for Time Use Statistics 2016}
\format{
A \link{tibble} with 230 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{https://unstats.un.org/unsd/demographic-social/time-use/icatus-2016/}
}
\usage{
icatus_2016
}
\description{
See \url{https://unstats.un.org/unsd/classifications/Family/Detail/2083}
}
\keyword{datasets}
